feat: background animation using ThreeJs
