+++ src/components/particule-background/index.ts
@@ -1,12 +1,6 @@
-﻿import * as THREE from 'three';
-import {camera, cameraTarget, initCamera, initCanvas, initScene, initStage, renderer, scene} from "./scene.ts";
+﻿import "./styles.css";
+import {animate, initCamera, initCanvas, initScene, initStage, renderer} from "./scene.ts";
 import {initBgObjects} from "./particules.ts";
-
-
-let
-    cameraLookAt = new THREE.Vector3(0, 0, 0);
-
-
 
 //-----------------------------------------------------------------------
 
@@ -20,19 +14,3 @@ export function animatedBackground() {
 
     return renderer.domElement;
 }
-
-//-----------------------------------------------------------------------
-
-
-const animate = () => {
-    requestAnimationFrame(animate);
-    camera.position.lerp(cameraTarget, 0.2);
-    camera.lookAt(cameraLookAt);
-    render();
-}
-
-const render = () => {
-    renderer.render(scene, camera);
-}
-//endregion
-
