+++ src/components/particule-background/styles.css
@@ -1,0 +1,3 @@
+﻿canvas {
+    background-color: #1D1D1D;
+}
