+++ src/index.ts
@@ -2,7 +2,6 @@ import './views/home-view/styles.css';
 import {buildVerticalNav} from "./components/vertical-nav";
 import {buildViewBase} from "./views/utils";
 import {buildProjectPage} from "./content/projects";
-import {animatedBackground, renderer} from "./components/particule-background";
 
 
 function init() {
@@ -10,11 +9,10 @@ function init() {
     const contentPage = buildViewBase();
     const verticalNav = buildVerticalNav();
 
-    //body.appendChild(contentPage);
-    //contentPage.appendChild(verticalNav);
-    body.appendChild(animatedBackground());
+    body.appendChild(contentPage);
+    contentPage.appendChild(verticalNav);
 
-    //buildProjectPage("space-compass");
+    buildProjectPage("space-compass");
 }
 
 init();
