+++ src/views/utils/styles.css
@@ -117,23 +117,14 @@
             background-image: url("./assets/background-test.png");
             width: 100%;
             height: 100%;
+            z-index: -11;
         }
 
         .layer-b {
-            position: fixed;
-            background-image: url("./assets/background-test-2.png");
-            transform: scaleX(-1);
-            width: 30%;
-            height: 100%;
+            position: relative;
+            z-index: -10;
         }
 
-        .layer-c {
-            position: fixed;
-            background-image: url("./assets/background-test-2.png");
-            right: 0;
-            width: 20%;
-            height: 100%;
-        }
     }
 }
